+++ package.json
@@ -29,7 +29,7 @@
   ],
   "main": "./dist/sema-ui.umd.js",
   "module": "./dist/sema-ui.es.js",
-  "types": "./src/index.d.ts",
+  "types": "./dist/index.d.ts",
   "homepage": "https://github.com/JEGDevStudios/Sema-UI",
   "repository": {
     "type": "git",
@@ -53,6 +53,9 @@
     "lit": "^3.0.0"
   },
   "devDependencies": {
-    "vite": "^7.3.1"
+    "tslib": "^2.8.1",
+    "typescript": "^5.9.3",
+    "vite": "^7.3.1",
+    "vite-plugin-dts": "^4.5.4"
   }
 }
