--- src/index.js
@@ -1,3 +1,0 @@
-export const version = "0.0.1";
-import { SemaButton } from "./components/SemaButton/SemaButton.js";
-import { SemaAlerts } from "./components/SemaAlerts/SemaAlerts.js";
