+++ src/index.ts
@@ -1,0 +1,3 @@
+export const version = "0.0.1";
+export * from "./components/SemaButton/SemaButton";
+export * from "./components/SemaAlerts/SemaAlerts";
