+++ vite.config.js
@@ -1,12 +1,11 @@
 import { defineConfig } from 'vite';
+import dts from 'vite-plugin-dts';
 
 export default defineConfig({
   build: {
     lib: {
-      entry: 'src/index.js',
-      
+      entry: 'src/index.ts',
       name: 'SemaUI',
-      
       fileName: (format) => `sema-ui.${format}.js`
     },
     rollupOptions: {
@@ -18,5 +17,11 @@ export default defineConfig({
         }
       }
     }
-  }
+  },
+  plugins: [
+    dts({ 
+      rollupTypes: true,
+      insertTypesEntry: true 
+    }) 
+  ]
 });
